# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk aplikasi manajemen kasir yang di-update dengan menggunakan database PostgreSQL. Terdapat penambahan endpoint untuk mengelola kategori produk serta relasi antara produk dan kategori.",
        "title": "Kasir API",
        "contact": {},
        "version": "1.0.1"
    },
    "paths": {
        "/api/checkout": {
            "post": {
                "description": "Melakukan checkout barang: format data yang harus diisi { items: [ { product_id, quantity } ]}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkout"
                ],
                "summary": "Checkout Product",
                "parameters": [
                    {
                        "description": "New Checkout Data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CheckoutRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create checkout",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/kategori": {
            "get": {
                "description": "Mengambil semua data kategori produk (Challange (Optional))",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category",
                    "produk"
                ],
                "summary": "Get All Categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Categories"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get categories",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/produk": {
            "get": {
                "description": "Mengambil semua data produk. Terdapat opsi untuk mendapatkan detail kategori produk",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produk"
                ],
                "summary": "Get All Products",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Tampilkan Detail Kategori Produk",
                        "name": "details",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tampilkan Detail Kategori Produk Berdasarkan Pencarian Nama",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get products",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Menambahkan data produk baru, data yang perlu diisi: { category_id, name, price, stock }",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produk"
                ],
                "summary": "Create New Product",
                "parameters": [
                    {
                        "description": "New Product Data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create product",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/produk/{id}": {
            "get": {
                "description": "Mengambil data produk berdasarkan ID. Terdapat opsi untuk mendapatkan detail kategori produk",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produk"
                ],
                "summary": "Get Product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Tampilkan Detail Kategori Produk",
                        "name": "details",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid product ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Memperbarui data produk berdasarkan ID, data yang dapat diubah: { category_id, name, price, stock }",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produk"
                ],
                "summary": "Update Product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Product Data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to update product",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus data produk berdasarkan ID",
                "tags": [
                    "produk"
                ],
                "summary": "Delete Product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid product ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to delete product",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/report": {
            "get": {
                "description": "Mengambil laporan data transaksi penjualan barang berdasarkan tanggal yang dipilih",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get Transaction Report By Selected Date",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2026-01-01",
                        "description": "Tanggal awal (Format: YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-02-01",
                        "description": "Tanggal akhir (Format: YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Report"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get report",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/report/hari-ini": {
            "get": {
                "description": "Mengambil laporan data transaksi penjualan barang khusus hari ini",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get Today's Transaction Report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Report"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get report",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Categories": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.CheckoutItem": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "models.CheckoutRequest": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CheckoutItem"
                    }
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "category_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "models.Report": {
            "type": "object",
            "properties": {
                "produk_terlaris": {
                    "type": "object",
                    "properties": {
                        "nama": {
                            "type": "string"
                        },
                        "qty_terjual": {
                            "type": "integer"
                        }
                    }
                },
                "total_revenue": {
                    "type": "number"
                },
                "total_transaksi": {
                    "type": "integer"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionDetail"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "integer"
                }
            }
        },
        "models.TransactionDetail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "subtotal": {
                    "type": "integer"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        }
    }
}